Fixes #2021 Move all refs to GoDiagram into UI project
