+++ src/MoBi.Presentation/Presenter/BaseDiagram/MoBiBaseDiagramPresenter.cs
@@ -1,14 +1,12 @@
 using System.Collections.Generic;
 using System.Drawing;
 using System.IO;
-using System.Linq;
 using MoBi.Assets;
 using MoBi.Core.Domain.Model;
 using MoBi.Core.Events;
 using MoBi.Core.Services;
 using MoBi.Presentation.Settings;
 using MoBi.Presentation.Views.BaseDiagram;
-using Northwoods.Go;
 using OSPSuite.Core;
 using OSPSuite.Core.Diagram;
 using OSPSuite.Core.Domain;
@@ -30,6 +28,7 @@ namespace MoBi.Presentation.Presenter.BaseDiagram
       //Model manipulation
       void ModelSelect(string id);
       void Link(IBaseNode node1, IBaseNode node2, object portObject1, object portObject2);
+      void Unlink(IBaseNode node1, IBaseNode node2, object portObject1, object portObject2);
 
       void SaveContainerToXml(IContainerBase containerBase, string diagramTemplateXmlFilePath);
 
@@ -41,7 +40,6 @@ namespace MoBi.Presentation.Presenter.BaseDiagram
       void ApplyLayoutTemplateToSelection();
 
       void ConfigureLayout();
-      void RemoveLinks(GoSelection links);
    }
 
    public interface IMoBiBaseDiagramPresenter<T> : IMoBiBaseDiagramPresenter, IBaseDiagramPresenter<T>,
@@ -143,7 +141,7 @@ namespace MoBi.Presentation.Presenter.BaseDiagram
       }
 
       public abstract void Link(IBaseNode node1, IBaseNode node2, object portObject1, object portObject2);
-      protected abstract void Unlink(IBaseNode node1, IBaseNode node2, object portObject1, object portObject2);
+      public abstract void Unlink(IBaseNode node1, IBaseNode node2, object portObject1, object portObject2);
 
       public override void ShowContextMenu(IBaseNode baseNode, Point popupLocation, PointF locationInDiagramView)
       {
@@ -240,29 +238,6 @@ namespace MoBi.Presentation.Presenter.BaseDiagram
       {
          var forceLayoutConfigurationPresenter = IoC.Resolve<IForceLayoutConfigurationPresenter>();
          forceLayoutConfigurationPresenter.Edit(LayoutConfiguration);
-      }
-
-      public void RemoveLinks(GoSelection links)
-      {
-         links.ToList().Each(link =>
-         {
-            if (isBaseLink(link) && isGoLink(link))
-               unlinkBaseNodes(link as IBaseLink, link as GoLink);
-            else if (link.IsAnImplementationOf<INeighborhoodNode>())
-               unlinkNeighborhoodNode(link);
-         });
-      }
-
-      private static bool isGoLink(GoObject goLink) => goLink is GoLink;
-
-      private static bool isBaseLink(GoObject baseLink) => baseLink is IBaseLink;
-
-      private void unlinkBaseNodes(IBaseLink baseLink, GoLink goLink) => Unlink(baseLink.GetFromNode(), baseLink.GetToNode(), goLink.FromPort.UserObject, goLink.ToPort.UserObject);
-
-      private void unlinkNeighborhoodNode(GoObject itemToDelete)
-      {
-         var neighborhoodNode = (INeighborhoodNode)itemToDelete;
-         Unlink(neighborhoodNode.FirstNeighbor, neighborhoodNode.SecondNeighbor, null, null);
       }
 
       public void Undo() => DiagramModel.Undo();
