+++ src/MoBi.Presentation/Presenter/ReactionDiagram/IReactionDiagramPresenter.cs
@@ -1,8 +1,6 @@
-using System.Collections.Generic;
 using MoBi.Core.Domain.Model;
 using MoBi.Presentation.Presenter.BaseDiagram;
 using MoBi.Presentation.Views.BaseDiagram;
-using Northwoods.Go;
 using OSPSuite.Core.Diagram;
 using OSPSuite.Core.Domain.Builder;
 using OSPSuite.Presentation.Presenters;
@@ -16,7 +14,6 @@ namespace MoBi.Presentation.Presenter.ReactionDiagram
       void AddMoleculeNode();
       void LayerLayout(IContainerBase containerBase);
       void LayerLayout();
-      void RemoveSelection(IReadOnlyList<GoObject> objectsToBeRemoved);
       void Select(ReactionBuilder reactionBuilder);
       bool IsReactionNode(IBaseNode node);
       bool DisplayEductsRight(IBaseNode node);
