+++ src/MoBi.UI/Views/BaseDiagram/MoBiBaseDiagramView.cs
@@ -1,13 +1,15 @@
 ﻿using System.ComponentModel;
-using OSPSuite.Utility.Extensions;
+using System.Linq;
 using MoBi.Presentation.Presenter.BaseDiagram;
 using MoBi.Presentation.Views.BaseDiagram;
 using Northwoods.Go;
 using OSPSuite.Core.Diagram;
+using OSPSuite.Presentation.Diagram.Elements;
 using OSPSuite.UI.Diagram.Elements;
 using OSPSuite.UI.Extensions;
 using OSPSuite.UI.Services;
 using OSPSuite.UI.Views.Diagram;
+using OSPSuite.Utility.Extensions;
 
 namespace MoBi.UI.Views.BaseDiagram
 {
@@ -51,9 +53,9 @@ namespace MoBi.UI.Views.BaseDiagram
 
       private void onLinkCreated(GoSelectionEventArgs e)
       {
-         var link = (NewLink) e.GoObject;
-         var node1 = (IBaseNode) link.FromNode;
-         var node2 = (IBaseNode) link.ToNode;
+         var link = (NewLink)e.GoObject;
+         var node1 = (IBaseNode)link.FromNode;
+         var node2 = (IBaseNode)link.ToNode;
          var portObject1 = link.FromPort.UserObject;
          var portObject2 = link.ToPort.UserObject;
 
@@ -63,9 +65,32 @@ namespace MoBi.UI.Views.BaseDiagram
 
       protected virtual void OnSelectionDeleting(CancelEventArgs e)
       {
-         _moBiDiagramPresenter.RemoveLinks(_goView.Selection);
+         removeLinks(_goView.Selection);
          e.Cancel = true;
       }
+
+      private void removeLinks(GoSelection links)
+      {
+         links.ToList().Each(link =>
+         {
+            if (isBaseLink(link) && isGoLink(link))
+               unlinkBaseNodes(link as IBaseLink, link as GoLink);
+            else if (link.IsAnImplementationOf<INeighborhoodNode>())
+               unlinkNeighborhoodNode(link);
+         });
+      }
+
+      private void unlinkBaseNodes(IBaseLink baseLink, GoLink goLink) => _moBiDiagramPresenter.Unlink(baseLink.GetFromNode(), baseLink.GetToNode(), goLink.FromPort.UserObject, goLink.ToPort.UserObject);
+
+      private void unlinkNeighborhoodNode(GoObject itemToDelete)
+      {
+         var neighborhoodNode = (INeighborhoodNode)itemToDelete;
+         _moBiDiagramPresenter.Unlink(neighborhoodNode.FirstNeighbor, neighborhoodNode.SecondNeighbor, null, null);
+      }
+
+      private static bool isGoLink(GoObject goLink) => goLink is GoLink;
+
+      private static bool isBaseLink(GoObject baseLink) => baseLink is IBaseLink;
 
       private void onBackgroundContextClicked(GoInputEventArgs e)
       {
