+++ src/MoBi.UI/Views/ReactionDiagram/ReactionDiagramView.cs
@@ -1,20 +1,20 @@
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
-using OSPSuite.UI;
-using OSPSuite.Utility.Extensions;
 using MoBi.Presentation.Presenter.ReactionDiagram;
 using MoBi.Presentation.Views.BaseDiagram;
+using MoBi.UI.Presenters;
 using MoBi.UI.Views.BaseDiagram;
 using Northwoods.Go;
 using OSPSuite.UI.Diagram.Elements;
 using OSPSuite.UI.Services;
+using OSPSuite.Utility.Extensions;
 
 namespace MoBi.UI.Views.ReactionDiagram
 {
    public class ReactionDiagramView : MoBiBaseDiagramView, IReactionDiagramView
    {
-      private IReactionDiagramPresenter _reactionDiagramPresenter;
+      private ReactionDiagramPresenter _reactionDiagramPresenter;
 
       public ReactionDiagramView(IImageListRetriever imageListRetriever)
          : base(imageListRetriever)
@@ -23,7 +23,7 @@ namespace MoBi.UI.Views.ReactionDiagram
 
       public void AttachPresenter(IReactionDiagramPresenter presenter)
       {
-         _reactionDiagramPresenter = presenter;
+         _reactionDiagramPresenter = presenter as ReactionDiagramPresenter;
          base.AttachPresenter(presenter);
       }
 
@@ -31,12 +31,12 @@ namespace MoBi.UI.Views.ReactionDiagram
       {
          var goObjects = _goView.Selection.ToList();
          // Any nodes that match our types and we process this delete
-         if(anyMoleculesOrReactions(goObjects))
+         if (anyMoleculesOrReactions(goObjects))
             _reactionDiagramPresenter.RemoveSelection(goObjects);
          // Otherwise pass requests to delete links and other items to the base class
          else
             base.OnSelectionDeleting(e);
-         
+
          e.Cancel = true;
       }
